+++ src/main/java/codesquad/aspect/SessionCheck.java
@@ -1,0 +1,12 @@
+package codesquad.aspect;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface SessionCheck {
+    String value() default "";
+}
